moved surface cache in own file
